+++ packages/core/solidity/src/account.ts
@@ -4,6 +4,7 @@ import { defineFunctions } from './utils/define-functions';
 import { printContract } from './print';
 import { defaults as commonDefaults, withCommonDefaults, type CommonOptions } from './common-options';
 import { setInfo } from './set-info';
+import { addSigner, signerFunctions, type SignerOptions } from './signer';
 
 export const defaults: Required<AccountOptions> = {
   ...commonDefaults,
@@ -16,9 +17,6 @@ export const defaults: Required<AccountOptions> = {
 
 export const accountBaseOptions = ['AccountCore', 'Account'] as const;
 export type AccountBaseOptions = (typeof accountBaseOptions)[number];
-
-export const SignerOptions = [false, 'ERC7702', 'ECDSA', 'P256', 'RSA'] as const;
-export type SignerOptions = (typeof SignerOptions)[number];
 
 export const ERC7579Options = [false, 'AccountERC7579', 'AccountERC7579Hooked'] as const;
 export type ERC7579Options = (typeof ERC7579Options)[number];
@@ -94,7 +92,7 @@ function addParents(c: ContractBuilder, opts: AccountOptions): void {
   // Extensions
   addERC7821(c, opts);
   addERC7579(c, opts);
-  addSigner(c, opts);
+  addSigner(c, opts.signer ?? false);
 }
 
 function addERC7821(c: ContractBuilder, opts: AccountOptions): void {
@@ -123,32 +121,6 @@ function addERC7579(c: ContractBuilder, opts: AccountOptions): void {
   c.addOverride({ name: 'AccountERC7579' }, functions._validateUserOp);
 }
 
-function addSigner(c: ContractBuilder, opts: AccountOptions): void {
-  if (!opts.signer) return;
-  c.addParent({
-    name: `Signer${opts.signer}`,
-    path: `@openzeppelin/community-contracts/contracts/utils/cryptography/Signer${opts.signer}.sol`,
-  });
-  c.addOverride({ name: `Signer${opts.signer}` }, functions._rawSignatureValidation);
-
-  // ERC-7702 doesn't require initialization
-  if (opts.signer === 'ERC7702') return;
-
-  c.addParent({
-    name: 'Initializable',
-    path: '@openzeppelin/contracts/proxy/utils/Initializable.sol',
-  });
-  const fn = functions[`initialize${opts.signer}`];
-  c.addModifier('initializer', fn);
-  c.addFunctionCode(
-    `_setSigner(${fn.args
-      .map(({ name }) => name)
-      .join(', ')
-      .trimEnd()});`,
-    fn,
-  );
-}
-
 function addEIP712(c: ContractBuilder, opts: AccountOptions): void {
   // EIP712 is only required for the Account and AccountERC7579 (both use EIP-712 as part of ERC-7739)
   if (opts.accountBase === 'Account' || opts.ERC7579) {
@@ -170,7 +142,7 @@ function overrideRawSignatureValidation(c: ContractBuilder, opts: AccountOptions
         name: 'AbstractSigner',
         path: '@openzeppelin/community-contracts/contracts/utils/cryptography/AbstractSigner.sol',
       });
-      c.addOverride({ name: 'AbstractSigner' }, functions._rawSignatureValidation);
+      c.addOverride({ name: 'AbstractSigner' }, signerFunctions._rawSignatureValidation);
       // If using `AccountERC7579Hooked`, the base contract won't be imported
       if (opts.ERC7579 === 'AccountERC7579Hooked') {
         c.addImportOnly({
@@ -178,14 +150,14 @@ function overrideRawSignatureValidation(c: ContractBuilder, opts: AccountOptions
           path: '@openzeppelin/community-contracts/contracts/account/extensions/AccountERC7579.sol',
         });
       }
-      c.addOverride({ name: 'AccountERC7579' }, functions._rawSignatureValidation);
+      c.addOverride({ name: 'AccountERC7579' }, signerFunctions._rawSignatureValidation);
       c.addFunctionCode(
         '// Force signer validation first, and fallback to ERC-7579',
-        functions._rawSignatureValidation,
+        signerFunctions._rawSignatureValidation,
       );
       c.addFunctionCode(
         `// return Signer${opts.signer}._rawSignatureValidation(hash, signature) || AccountERC7579._rawSignatureValidation(hash, signature);`,
-        functions._rawSignatureValidation,
+        signerFunctions._rawSignatureValidation,
       );
     } else {
       // Signer implements _rawSignatureValidation alone
@@ -195,41 +167,14 @@ function overrideRawSignatureValidation(c: ContractBuilder, opts: AccountOptions
       // AccountERC7579 implements _rawSignatureValidation alone
     } else {
       // Custom validation logic
-      c.addOverride({ name: opts.accountBase }, functions._rawSignatureValidation);
-      c.setFunctionBody(['// Custom validation logic', 'return false;'], functions._rawSignatureValidation);
+      c.addOverride({ name: opts.accountBase }, signerFunctions._rawSignatureValidation);
+      c.setFunctionBody(['// Custom validation logic', 'return false;'], signerFunctions._rawSignatureValidation);
     }
   }
 }
 
 const functions = {
   ...defineFunctions({
-    initializeECDSA: {
-      kind: 'public' as const,
-      args: [{ name: 'signer', type: 'address' }],
-    },
-    initializeP256: {
-      kind: 'public' as const,
-      args: [
-        { name: 'qx', type: 'bytes32' },
-        { name: 'qy', type: 'bytes32' },
-      ],
-    },
-    initializeRSA: {
-      kind: 'public' as const,
-      args: [
-        { name: 'e', type: 'bytes memory' },
-        { name: 'n', type: 'bytes memory' },
-      ],
-    },
-    _rawSignatureValidation: {
-      kind: 'internal' as const,
-      args: [
-        { name: 'hash', type: 'bytes32' },
-        { name: 'signature', type: 'bytes calldata' },
-      ],
-      returns: ['bool'],
-      mutability: 'view' as const,
-    },
     _signableUserOpHash: {
       kind: 'internal' as const,
       args: [