+++ packages/core/solidity/src/build-generic.ts
@@ -13,6 +13,8 @@ import { buildGovernor } from './governor';
 import type { Contract } from './contract';
 import { buildAccount } from './account';
 import type { AccountOptions } from './account';
+import { buildPaymaster } from './paymaster';
+import type { PaymasterOptions } from './paymaster';
 
 export interface KindedOptions {
   ERC20: { kind: 'ERC20' } & ERC20Options;
@@ -21,6 +23,7 @@ export interface KindedOptions {
   Stablecoin: { kind: 'Stablecoin' } & StablecoinOptions;
   RealWorldAsset: { kind: 'RealWorldAsset' } & StablecoinOptions;
   Account: { kind: 'Account' } & AccountOptions;
+  Paymaster: { kind: 'Paymaster' } & PaymasterOptions;
   Governor: { kind: 'Governor' } & GovernorOptions;
   Custom: { kind: 'Custom' } & CustomOptions;
 }
@@ -46,6 +49,9 @@ export function buildGeneric(opts: GenericOptions): Contract {
 
     case 'Account':
       return buildAccount(opts);
+
+    case 'Paymaster':
+      return buildPaymaster(opts);
 
     case 'Governor':
       return buildGovernor(opts);