+++ packages/core/solidity/src/kind.ts
@@ -20,6 +20,7 @@ function isKind<T>(value: Kind | T): value is Kind {
     case 'Stablecoin':
     case 'RealWorldAsset':
     case 'Account':
+    case 'Paymaster':
     case 'Governor':
     case 'Custom':
       return true;