+++ packages/core/solidity/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.5.4 (XXXX-XX-XX)
 
-- Add `account` contract type.
+- Add `account` and `paymaster` contract types for ERC-4337.
 
 ## 0.5.3 (2025-03-13)